Crate mine

Minesweeper game

This is my implementation of the classic minesweeper game.

This particular program can run as a cli, as a tui or as a gui.

Modules

cells 🔒

Cell types and utilities for the Mine Sweeper game.

cli 🔒

Cli module of the minesweeper game.

game_play 🔒

Main game logic and board implementation.

gui 🔒

GUI implementation of the game using Yew.

tui 🔒

TUI implementation of the game.

Functions

cli_args 🔒

Handles the command line interface arguments.

main 🔒

Main entrypoint of the program

start_cli 🔒

Starts the command line interface

start_gui 🔒

Starts the graphical user interface

start_tui 🔒

Starts the terminal user interface