mine::gui

Module enabled

When the mine_gui is enabled we are going to want to run this code instead

Structs

App

The main Yew functional component for the Minesweeper GUI.

Functions

run_gui

Bootstrap function called when the WASM module is instantiated.